added correct databinding
