+++ App/automotive/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
     <uses-feature
         android:name="android.hardware.type.automotive"
         android:required="true" />
-
+    <uses-permission android:name="android.car.permission.CAR_IDENTIFICATION"  />
     <application
         android:allowBackup="true"
         android:appCategory="audio"
