+++ App/automotive/src/main/java/com/amos/infotaimos/view/VehiclePropertiesPage.kt
@@ -1,5 +1,7 @@
 package com.amos.infotaimos.view
 
+import android.os.Bundle
+import android.view.View
 import androidx.fragment.app.viewModels
 import com.amos.infotaimos.ViewBindingFragment
 import com.amos.infotaimos.databinding.FragmentVehiclePropertiesPageBinding
@@ -7,4 +9,15 @@ import com.amos.infotaimos.viewmodel.VehiclePropertiesPageViewModel
 
 class VehiclePropertiesPage : ViewBindingFragment<FragmentVehiclePropertiesPageBinding>() {
     private val viewModel: VehiclePropertiesPageViewModel by viewModels()
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        context?.let { viewModel.registerVinCallback(it) }
+        viewModel.vin.observe(viewLifecycleOwner) { vin ->
+            binding.vinTile.tileVinText.text = vin
+        }
+    }
+
+
 }
