--- App/automotive/src/main/java/com/amos/infotaimos/viewmodel/VinTileViewModel.kt
@@ -1,6 +1,0 @@
-package com.amos.infotaimos.viewmodel
-
-import androidx.lifecycle.ViewModel
-
-class VinTileViewModel : ViewModel(){
-}
