+++ App/automotive/src/main/res/layout/tile_vin.xml
@@ -3,15 +3,13 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    xmlns:tools="http://schemas.android.com/tools">
+    android:layout_height="wrap_content">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:padding="8dp"
-        android:background="@color/landing_green"
-        tools:context=".view.VinTile">
+        android:background="@color/landing_green">
 
         <ImageView
             android:id="@+id/tile_vin_icon"
